rotaryencoder: Improve rotary encoder behaviour.
Increase the priority of the foreground task so that
when interrupts fire, we handle them quickly enough to
mostly never miss the values.

Catch rising edges as well as falling, so we don't miss
code values.

Ignore spurious wakeups where the pins haven't changed,
since reading the ADC fires our GPIO interrupt for some
reason not yet understood.

Reverse the direction of the rotary encoder - clockwise is
increment, and anti-clockwise is decrement.